API 날씨 데이터 따라 UI 변경

+++ app/src/main/java/com/udemy/weatherapp/model/Sys.kt
@@ -3,5 +3,5 @@ package com.udemy.weatherapp.model
 import java.io.Serializable
 
 data class Sys(
-    val type: Int, val message: Double, val country: String, val sunrise: Int, val sunset: Int
+    val type: Int, val message: Double, val country: String, val sunrise: Long, val sunset: Long
 ) : Serializable
